requirements.txt
pandas
pdfplumber
PyPDF2
openpyxl
xlsxwriter
configparser
ttkbootstrap
numpy
Pillow
